rename: entities의 tags 폴더 이동
