+++ src/entities/shared-game/ui/tag/reset-button.stories.ts
