+++ src/entities/shared-game/ui/tag/tag.stories.ts
