+++ src/entities/shared-game/ui/tag/tag.tsx
